+++ CHANGELOG.md
@@ -5,9 +5,16 @@ PrestaShop 1.6.x and 1.7.x
 
 # Changelog
 
-## [3.5.8] - 2023-08-04
+
+## [3.5.9] - 2023-08-04
 ### Added
 - Set the **checkout** design of the Credit Card form by default for new installations.
+
+## [3.5.8] - 2023-07-18
+### Added
+- Make Apple Pay compatible with **The Checkout** (One Page Checkout module)
+### Fixed
+- Fix: Resetting payment status on late Ok or Notification Callback.
 
 ## [3.5.7] - 2023-07-05
 ### Added